Ignore false positive on golintci
The import cannot be 'import "embed"' because it is removed by
gopls in VSCode. Using import _ "embed" avoid that situation,
but complain on the linter.

Considering this as a false positive, I try to ignore this
be ignoring the import line for "embed" which is necessary
to embed a file in the 'templateRepomdXmls' variable.

Signed-off-by: Alejandro Visiedo <avisiedo@redhat.com>