Fixes 1841: correct param validate mock (#286)
The param validator mock was improperly defined, and thus
improperly used.  When replaced with a proper mock, it caused
a hang, this due to the wait group within the handler not properly
marking Done() if an unexpected error occured.  By using defer, we can
mark it Done() no matter what happens within the go routine.