Refactor: move helpers and fixtures to shared dir
This moves and in some cases slightly polishes helpers, fixtures and the
API client to a new directory which is going to contain any shared test
utilities between our projects (through git submodules and sparse
checkout).
Along with this a new `package.json` file was added to the to be shared
directory which should contain any dependencies required by these
utilities. This shared utilities library should then be included in
`package.json` files of repos consuming it and also to their TS configs
for compilation.