HMS-5913: add package dates from RHEL lifecycle
Adds package streams info to package_sources. Pulls package stream info
from the roadmap API and adds it to package sources if the RPM name
matches.

Also adds RHEL EoL as the end date for any package source that is not in
the roadmap API. Uses the end date for the latest released version of
the matching RHEL major version.