Adding the code block again to list and delete repo
As there were changes in action file there was no need to keep this code
block. But to be sure that if QE or anyone wants to run it locally we
need to ensure that the pop repo is not present and if it is in custom
repo we should delete it.

So, adding that block again to ensure proper test cleanup.