add: option for resumable and non-resumable uploads
In the /repositories/uploads/ api if there has been already a previous
upload matching the sha256 and chunk_size created, it was used on
subsequent uploads (for instance when the full upload process had been
interrupted and was resumed later) and the existing UUID was returned.
There was no option to resume with a fresh upload, returning a new UUID.
This adds the new option to allow non-resumable uploads, passing
resumable=false parameter in the post request body for the api. It makes it
easier for users that use this api with their custom scripts, not having
to write resumable logic. Passing resumable=true will behave as it did,
returning the existing UUID.