update: pulp calls return 200 with no task for update calls
Previously pulp always created an internal pulp task and returned the task href,
when there had been triggered an update call to domains or remote, irrespective if
the update was changing anything or not. From now on, when an update call to
domain (UpdateDomainIfNeeded) or remote (UpdateRmpRemote) does not change anything,
pulp does not create a task. This is indicated by 200 status code. If there is an actual
change, pulp sends 202 with taskHref back. This commit updates the parts of code
that deal with these pulp functions to take this pulp api change into account.