test: improve stability of locally running PW tests
These changes aim to improve the stability of Playwright test when they
are being run locally.
Switches repository introspection to use a random repo URL, since it
doesn't require a specific one.
Modifies the CRUD org ID test to use a specific user, since it checks
for a concrete value of the org ID.